requirements.txt
bokeh==2.4.1
matplotlib==3.5.1
numpy==1.23.4
pandas==1.4.1
Pillow==9.1.0
scipy==1.8.0
seaborn~=0.12.1
streamlit==1.13.0
xlrd < 2.0.1
openpyxl < 3.0.7
pi
streamlit-authenticator
PyYAML~=6.0